+++ internal/controller/federation_controller.go
@@ -22,6 +22,7 @@ import (
 	"time"
 
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/api/meta"
 	"k8s.io/apimachinery/pkg/runtime"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/builder"
@@ -214,6 +215,15 @@ func (r *FederationReconciler) Reconcile(
 	// Handle deletion of the federation resource
 	if !fed.GetDeletionTimestamp().IsZero() {
 		if isGuest {
+			hasDependents, depErr := r.resourcesDependOnFederation(ctx, fed.Namespace, fed.Status.FederationContextId)
+			if depErr != nil {
+				log.Error(depErr, ">>> [Federation] Error checking for dependent resources before deletion.", "name", fed.Name, "namespace", fed.Namespace)
+				return ctrl.Result{}, depErr
+			}
+			if hasDependents {
+				log.Info(">>> [Federation] BLOCKED deletion: dependent resources still reference this federationContextId.", "name", fed.Name, "namespace", fed.Namespace, "federationContextId", fed.Status.FederationContextId)
+				return ctrl.Result{RequeueAfter: 15 * time.Second}, nil
+			}
 			if err := extClient.DeleteFederation(ctx, &fed); err != nil {
 				log.Error(err, ">>> [Federation] Error during the deletion.", "name", fed.Name, "namespace", fed.Namespace)
 				return ctrl.Result{}, err
@@ -356,6 +366,51 @@ func (r *FederationReconciler) Reconcile(
 		}
 	}
 	return ctrl.Result{}, nil
+}
+
+func (r *FederationReconciler) resourcesDependOnFederation(ctx context.Context, namespace string, federationContextId string) (bool, error) {
+	if federationContextId == "" {
+		return false, nil
+	}
+
+	dependentLists := []client.ObjectList{
+		&v1beta1.ImageList{},
+		&v1beta1.AvailabilityZoneList{},
+		&v1beta1.ArtefactList{},
+		&v1beta1.ApplicationDeploymentList{},
+		&v1beta1.ApplicationOnboardingList{},
+	}
+
+	for _, list := range dependentLists {
+		if err := r.List(ctx, list, client.InNamespace(namespace)); err != nil {
+			return false, err
+		}
+		objs, err := meta.ExtractList(list)
+		if err != nil {
+			return false, err
+		}
+		for _, obj := range objs {
+			var fedContextId string
+			switch cr := obj.(type) {
+			case *v1beta1.Image:
+				fedContextId = cr.Spec.FederationContextId
+			case *v1beta1.AvailabilityZone:
+				fedContextId = cr.Spec.FederationContextId
+			case *v1beta1.Artefact:
+				fedContextId = cr.Spec.FederationContextId
+			case *v1beta1.ApplicationDeployment:
+				fedContextId = cr.Spec.FederationContextId
+			case *v1beta1.ApplicationOnboarding:
+				fedContextId = cr.Spec.FederationContextId
+			default:
+				continue
+			}
+			if fedContextId == federationContextId {
+				return true, nil
+			}
+		}
+	}
+	return false, nil
 }
 
 func (r *FederationReconciler) guestFederationActions(ctx context.Context, fed *v1beta1.Federation, isRest bool, extClient ExternalFederationClient, annotations map[string]string) error {
